requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
azure-cosmos==4.7.0
azure-search-documents==11.4.0
azure-identity==1.15.0
ragas==0.1.9
langchain==0.1.0
openai==1.3.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
